Move metrics types to Contracts to break Gateway→Core leak (#154)
`SessionLifecycleMetrics` and `ThrottleMetrics` are pure value-type
counters consumed by the FIXP session via `FixpSessionOptions`. They
lived in `B3.Exchange.Core` which forced any `FixpSessionOptions`
caller (and the Gateway itself) to depend on Core types just to wire
metrics — a downward dependency that fights the intended ports/
adapters layout.

This commit:
- Adds `SessionLifecycleMetrics` and `ThrottleMetrics` to
  `B3.Exchange.Contracts` (the canonical port-types assembly).
- Removes the duplicates from `B3.Exchange.Core/MetricsRegistry.cs`;
  `MetricsRegistry` now imports them via `using B3.Exchange.Contracts`
  and exposes them through the same `Sessions`/`Throttle` properties.
- Updates `FixpSessionOptions` to type the optional metrics fields as
  `B3.Exchange.Contracts.*`.
- Updates `FixpSessionThrottleTests` constructions to match.

The Gateway still keeps a project reference to Core for the legacy
`IInboundCommandSink` / `IGatewayResponseChannel` ports — fully
removing the project reference is parked for the larger Wave C
"Contracts canonical" cleanup (issue C1 in the roadmap).

Closes #154
